LPS-104369 Moving to seo.canonical.url package
