+++ modules/apps/layout/layout-seo-api/src/main/java/com/liferay/layout/seo/canonical/url/LayoutSEOCanonicalURLProvider.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.layout.seo;
+package com.liferay.layout.seo.canonical.url;
 
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.model.Layout;
