+++ modules/apps/layout/layout-seo-api/src/main/resources/com/liferay/layout/seo/canonical/url/packageinfo
